Add RestrictedLocalFiniteElementView wrapper
RestrictedLocalFiniteElementView wrappes a given LocalFiniteElement
and can either be active, i.e. everyhing behaves just like the
underlying LocalFiniteElement, or inactive, i.e. there are no shape
functions present and `size() == 0`.

The retional behind this wrapper is to restrict the support of a
global basis to a subdomain and in this course disable a
LocalFiniteElement on elements outside of the support.