[bug][draft] Check traces of dynamic order Lagrange basis for pyramid
This is a specific test for generic FE implementation of
the Lagrange basis functions on the quadrilateral pyramid.
The test checks that an order one basis function vanishing
at all vertices of a face also vanishes at its center.
If the traces are in the appropriate polynomial spaces,
this should be true, but the check currently fails.