Added support for SIMD interpolation including updated tests
Now it is possible to interpolate into SIMD type coefficient vector while
keeping the used basis scalar. Therefore, inside the localinterpolation
some types based on the basis had to replaced by types independent of the
basis.

Currently, this is not supported by the VirtualInterface. There is some work
to be done.

During the tests (test-fe.hh and test-localfe.hh) now additionally the
SIMD interpolation is tested - unless for virtual Finite Elements.